+++ src/main/java/plant/Peach.java
@@ -6,7 +6,7 @@ public class Peach extends Tree {
     }
 
     @Override
-    public void water(double amount) {
+    public void water(int amount) {
         System.out.println("Watering the peach tree with " + amount + " units of water.");
     }
 }
