+++ src/main/java/plant/Plant.java
@@ -6,13 +6,13 @@ import java.util.List;
 
 public abstract class Plant{
     private String name;
-    private double currentWaterLevel;
+    private int currentWaterLevel;
     private boolean isUnderAttack;  // Flag to indicate if the plant is being attacked
     private WaterSensor waterSensor;
     private PlantType plantType;
     private int health;  // Health of the plant
 
-    public Plant(String name, PlantType plantType, double currentWaterLevel) {
+    public Plant(String name, PlantType plantType, int currentWaterLevel) {
         this.name = name;
         this.plantType = plantType;
         this.currentWaterLevel = currentWaterLevel;
@@ -23,12 +23,12 @@ public abstract class Plant{
 
     // Getter
     public String getName() { return this.name; }
-    public double getMinWaterLevel() { return this.plantType.getMinWaterLevel(); }
-    public double getMaxWaterLevel() { return this.plantType.getMaxWaterLevel(); }
-    public double getMinTemperatureLevel() { return this.plantType.getMinTemperature(); }
-    public double getMaxTemperatureLevel() { return this.plantType.getMaxTemperature(); }
-    public double getCurrentWaterLevel() { return this.currentWaterLevel; }
-    public double getPlantLowWaterThreshold() {return this.plantType.getLowWaterThreshold();}
+    public int getMinWaterLevel() { return this.plantType.getMinWaterLevel(); }
+    public int getMaxWaterLevel() { return this.plantType.getMaxWaterLevel(); }
+    public int getMinTemperatureLevel() { return this.plantType.getMinTemperature(); }
+    public int getMaxTemperatureLevel() { return this.plantType.getMaxTemperature(); }
+    public int getCurrentWaterLevel() { return this.currentWaterLevel; }
+    public int getPlantLowWaterThreshold() {return this.plantType.getLowWaterThreshold();}
     public List<String> getPestList(){ return this.plantType.getPests();}
     public boolean isUnderAttack() { return isUnderAttack; }
     public int getHealth() { return health; }
@@ -36,12 +36,12 @@ public abstract class Plant{
 
     // Setter
     public void setHealth(int health){ this.health = health;}
-    public void setCurrentWaterLevel(double currentWaterLevel) { this.currentWaterLevel = currentWaterLevel;}
+    public void setCurrentWaterLevel(int currentWaterLevel) { this.currentWaterLevel = currentWaterLevel;}
 
     /**
      * Update water level
      */
-    public void updateWaterLevel(double newWaterLevel){
+    public void updateWaterLevel(int newWaterLevel){
         this.waterSensor.updateWaterLevel(newWaterLevel);
     }
 
@@ -54,5 +54,5 @@ public abstract class Plant{
     /**
      * Abstract methods to be implemented by subclasses
      */
-    public abstract void water(double amount);
+    public abstract void water(int amount);
 }
